requirements.txt
Django==1.4
wsgiref==0.1.2


# add ons
django-taggit==0.9.3

